+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.1 (2019-03-??)
+
+### Fixes
+
+- Minor, aspect ratio __toString(), return with separator from constructor instead of default.
+- Minor, use alias for psr interfaces in logger and cache
+
 ## 1.0.0 (2019-03-12)
 
 ### Improved